// src/app/@core/models/workpermit.ts
export class Workpermit {
  ref_number: any;
  safety_or_concerns: any;
  safety_access: any;
  supervisor_review: any;
  supervisor_approval: any;
  hod_review: any;
  hod_approval: any;
  safety_officer_review: any;
  safety_officer_approval: any;
  scope: any;
  department: any;
  supervisor: any;
  project_manager: any;
  hod: any;
  safety_officer: any;
  permits: any;
  communication_plan: any;
}